+++ src/main/java/id/naturalsmp/naturalSchool/ui/UIManager.java
@@ -24,7 +24,7 @@ import java.util.logging.Level;
 
 public class UIManager {
 
-    public static final String GUI_VERSION = "1.7.2";
+    public static final String GUI_VERSION = "1.7.3";
 
     private static final MiniMessage MM = MiniMessage.miniMessage();